use trait elsewhere
