+++ crates/runtime/src/dataupdate.rs
@@ -26,6 +26,7 @@ use datafusion::physical_plan::{
     DisplayAs, DisplayFormatType, ExecutionPlan, Partitioning, PlanProperties,
 };
 use futures::TryStreamExt;
+use runtime_datafusion::query_engine::{DataUpdate, UpdateType};
 use tokio::sync::{Mutex, RwLock, broadcast};
 
 use datafusion::sql::TableReference;
@@ -167,24 +168,6 @@ impl DataUpdateBroadcaster {
 }
 
 use crate::datafusion::error::find_datafusion_root;
-
-#[derive(Debug, Clone, PartialEq)]
-pub enum UpdateType {
-    Append,
-    Overwrite,
-    Changes,
-}
-
-#[derive(Debug, Clone)]
-pub struct DataUpdate {
-    pub schema: SchemaRef,
-    pub data: Vec<RecordBatch>,
-    /// The type of update to perform.
-    /// If `UpdateType::Append`, the runtime will append the data to the existing dataset.
-    /// If `UpdateType::Overwrite`, the runtime will overwrite the existing data with the new data.
-    /// If `UpdateType::Changes`, the runtime will apply the changes to the existing data.
-    pub update_type: UpdateType,
-}
 
 pub struct StreamingDataUpdate {
     pub data: SendableRecordBatchStream,
