+++ crates/runtime/src/tools/builtin/sample/top_samples.rs
@@ -106,11 +106,7 @@ impl SampleFrom for TopSamplesParams {
                     tbl = tbl_quoted,
                 )))
                 .await?;
-
-            stream
-                .try_collect::<Vec<RecordBatch>>()
-                .await
-                .map_err(|e| -> Box<dyn std::error::Error + Send + Sync> { Box::new(e) })
+            stream.try_collect::<Vec<RecordBatch>>().await.boxed()
         }
         .instrument(current_span)
         .await?;
